Extract GenerateEmbeddings Controller & Refactor Folder Structure
