+++ utils/invalid_input.js
@@ -1,0 +1,6 @@
+function sendInvalidInputMessage(res, message) {
+    res.throw(422, message);
+    return false;
+}
+
+exports.sendInvalidInputMessage = sendInvalidInputMessage;
